Address feedback from Mark
