+++ ui/apps/platform/src/services/RolesService.ts
@@ -1,6 +1,6 @@
+import { Traits } from 'types/traits.proto';
 import axios from './instance';
 import { Empty } from './types';
-import { traitsOriginLabels } from '../constants/accessControl';
 
 const resourcesUrl = '/v1/resources';
 
@@ -80,14 +80,6 @@ export function fetchUserRolePermissions(): Promise<{ response: Role }> {
 
 const permissionSetsUrl = '/v1/permissionsets';
 
-export type Traits = {
-    mutabilityMode: MutabilityMode;
-    origin?: Origin;
-};
-
-export type MutabilityMode = 'ALLOW_MUTATE' | 'ALLOW_MUTATE_FORCED';
-export type Origin = 'IMPERATIVE' | 'DECLARATIVE' | 'DEFAULT';
-
 export type PermissionSet = {
     id: string;
     name: string;
@@ -95,16 +87,6 @@ export type PermissionSet = {
     resourceToAccess: PermissionsMap;
     traits?: Traits;
 };
-
-export function getTraitsOriginLabel(traits?: Traits): string {
-    return traits && traits.origin && traitsOriginLabels[traits.origin]
-        ? traitsOriginLabels[traits.origin]
-        : 'User';
-}
-
-export function isUserResource(traits?: Traits): boolean {
-    return traits == null || traits.origin === 'IMPERATIVE';
-}
 
 /*
  * Fetch entities and return array of objects.
