+++ ui/apps/platform/src/types/traits.proto.ts
@@ -1,0 +1,7 @@
+export type Traits = {
+    mutabilityMode: TraitsMutabilityMode;
+    origin?: TraitsOrigin;
+};
+
+export type TraitsMutabilityMode = 'ALLOW_MUTATE' | 'ALLOW_MUTATE_FORCED';
+export type TraitsOrigin = 'IMPERATIVE' | 'DECLARATIVE' | 'DEFAULT';
